requirements.txt
torch
numpy
pandas
PyYAML
scikit-learn
lyricsgenius
transformers